Assignment1/ex1_code/requirements.txt
numpy
matplotlib
scipy
pillow
opencv-python